build: instalação do react-router-dom; refactor: alteração de caminhos dos arquivos
