+++ encomendas/src/main/java/chegouAqui/DatabaseNamePrinter.java
@@ -1,4 +1,4 @@
-package com.chegouAqui.encomendas;
+package chegouAqui;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
