+++ encomendas/src/main/java/chegouAqui/EncomendasApplication.java
@@ -1,4 +1,4 @@
-package com.chegouAqui.encomendas;
+package chegouAqui;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
