+++ encomendas/src/main/java/chegouAqui/dto/LoginDTO.java
@@ -1,0 +1,20 @@
+package chegouAqui.dto;
+
+public class LoginDTO {
+    private String email;
+    private String senha;
+
+    public String getEmail(){
+        return email;
+    }
+    public void setEmail(String email){
+        this.email = email;
+    }
+    public String getSenha(){
+        return senha;
+    }
+    public void setSenha(String senha){
+        this.senha = senha;
+    }
+       
+}
