+++ encomendas/src/main/java/chegouAqui/repository/AdministradorRepository.java
@@ -1,6 +1,6 @@
-package com.chegouAqui.encomendas.repository;
+package chegouAqui.repository;
 
-import com.chegouAqui.encomendas.entity.Administrador;
+import chegouAqui.entity.Administrador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdministradorRepository extends JpaRepository<Administrador, Long> {
