+++ encomendas/src/main/java/chegouAqui/repository/EncomendasRepository.java
@@ -1,6 +1,6 @@
-package com.chegouAqui.encomendas.repository;
+package chegouAqui.repository;
 
-import com.chegouAqui.encomendas.entity.Encomendas;
+import chegouAqui.entity.Encomendas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EncomendasRepository extends JpaRepository<Encomendas, Long> {
