+++ encomendas/src/main/java/chegouAqui/repository/MoradorRepository.java
@@ -1,6 +1,6 @@
-package com.chegouAqui.encomendas.repository;
+package chegouAqui.repository;
 
-import com.chegouAqui.encomendas.entity.Morador;
+import chegouAqui.entity.Morador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MoradorRepository extends JpaRepository <Morador, Long> {
