+++ front/front-encomendas/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "axios": "^1.13.2",
     "react": "^19.1.1",
-    "react-dom": "^19.1.1"
+    "react-dom": "^19.1.1",
+    "react-router-dom": "^7.13.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.36.0",
